Build error fix after r14283.
BUG=NONE

Review URL: https://codereview.chromium.org/14060010

Patch from Marja Hölttä <marja@chromium.org>.

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@14286 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/heap-snapshot-generator.cc
@@ -2001,7 +2001,7 @@ void NativeObjectsExplorer::FillImplicitReferences() {
 
   int representative_objects_index = 0;
   UniqueId current_group_id(0);
-  size_t current_group_start = 0;
+  int current_group_start = 0;
   for (int i = 0; i <= ref_groups->length(); ++i) {
     if (i == 0)
       current_group_id = ref_groups->at(i).id;
